+++ src/components/UiCheckboxGroup/index.js
@@ -1,0 +1,4 @@
+import factoryComponentPlugin from '@/utils/factoryComponentPlugin';
+import UiCheckboxGroup from './UiCheckboxGroup';
+
+export default factoryComponentPlugin(UiCheckboxGroup);