+++ src/components/UiRadioGroup/index.js
@@ -1,0 +1,4 @@
+import factoryComponentPlugin from '@/utils/factoryComponentPlugin';
+import UiRadioGroup from './UiRadioGroup';
+
+export default factoryComponentPlugin(UiRadioGroup);